app/requirements.txt
requests
beautifulsoup4
pandas
geopandas
shapely
python-dotenv
transformers
torch
sentencepiece
scikit-learn
umap-learn
hdbscan
spacy
streamlit
fastapi
uvicorn
joblib
pyproj
folium
matplotlib
jupyterlab
pytest